Move CorruptionException to snappy package
To be replaced with MalformedInputException later
