+++ src/main/java/io/airlift/compress/snappy/CorruptionException.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.airlift.compress;
+package io.airlift.compress.snappy;
 
 public class CorruptionException extends RuntimeException
 {
